added: console helper to help interface with the progress bar

+++ sar/Commands/FileDestory.cs
@@ -27,43 +27,49 @@ namespace skylib.sar
 			{
 				throw new ArgumentException("incorrect number of arguments");
 			}
-			
-			string filePattern = args[1];
-			
+						
+			string filePattern = args[1];						
 			string root = Directory.GetCurrentDirectory();
+			IO.CheckRootAndPattern(ref root, ref filePattern);
 			List<string> files = IO.GetAllFiles(root, filePattern);
+						
+			ConsoleHelper.WriteLine(files.Count.ToString() + " file" + ((files.Count != 1) ? "s" : "") + " found");
 			
-			foreach (string file in files)
+			int count = 0;
+			
+			if (files.Count > 0)
 			{
-				try
+				if (Program.NoWarning || ConsoleHelper.Confirm("Destroy files? (y/n)"))
 				{
-					File.SetAttributes(file, FileAttributes.Normal);
-					StreamWriter sw = new StreamWriter(file);
-					sw.WriteLine("file corrupt");
-					sw.Close();
-					File.Delete(file);
-					Console.ForegroundColor = ConsoleColor.Cyan;
-					Console.Write("passed: ");
-					Console.ResetColor();
-					Console.WriteLine(file.Substring(root.Length + 1));
-				}
-				catch (Exception ex)
-				{
-					Console.ForegroundColor = ConsoleColor.Red;
-					Console.Write("failed: ");
-					Console.ResetColor();
-					Console.WriteLine(file.Substring(root.Length + 1));
-					
-					Console.ForegroundColor = ConsoleColor.Green;
-					Console.WriteLine("error: " + ex.Message);
-					Console.ResetColor();
+					foreach (string file in files)
+					{
+						try
+						{
+							File.SetAttributes(file, FileAttributes.Normal);
+							StreamWriter sw = new StreamWriter(file);
+							sw.WriteLine("file corrupt");
+							sw.Close();
+							File.Delete(file);
+							ConsoleHelper.Write("destroyed: ", ConsoleColor.Cyan);
+							ConsoleHelper.WriteLine(file.Substring(root.Length + 1));
+							count++;
+						}
+						catch (Exception ex)
+						{
+							ConsoleHelper.Write("failed: ", ConsoleColor.Red);
+							ConsoleHelper.WriteLine(file.Substring(root.Length + 1));
+							
+							if (Program.Debug)
+							{
+								ConsoleHelper.WriteException(ex);
+							}
+						}
+					}
 				}
 			}
-
-			Console.ForegroundColor = ConsoleColor.DarkYellow;
-			Console.WriteLine("Files Found " + files.Count.ToString() + " file" + ((files.Count != 1) ? "s" : ""));
-			Console.ResetColor();
 			
+			
+			ConsoleHelper.WriteLine("Files destroyed: " + count.ToString() + " of " + files.Count.ToString());
 			return Program.EXIT_OK;
 		}
 	}
